Add feature to create a new comment
